Lab is almost complete, minus a bug and commenting
The lab is almost complete. However, there is one major bug and there is no documentation.
The bug is caused by the deletion of dead hares. When a dead hare is deleted, the next hare takes its index in the ArrayList. Since the current deletion process relies on indexes, this causes some dead hares to not be removed from simulation.